+++ CHANGELOG.md
@@ -1,9 +1,14 @@
 # Changelog
 
+## Unreleased
+- Set mlflow experiment name as `aim.Run`'s experiment and parse the mlflow run name (mihran113)
+- Try to parse "string-ified" params values (sirykd)
+
 ## 0.1.1 
 - Fix MLFlow active runs' query (mihran113)
 - Add `PyPI` long description (mihran113)
 - Rename `convert` command to `sync` (mihran113)
+
 ## 0.1.0 
 
 - Add mlflow to aim convert command (mihran113)